v0.3.3
Implement jwt auth;
Refactor packages structure;
Configure CORS;
Implement access control by user role;

+++ src/main/java/com/sony/engineering/portalcadastro/controller/UserController.java
@@ -2,19 +2,14 @@ package com.sony.engineering.portalcadastro.controller;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.sony.engineering.portalcadastro.model.User;
 import com.sony.engineering.portalcadastro.service.UserService;
@@ -23,6 +18,9 @@ import com.sony.engineering.portalcadastro.service.UserService;
 public class UserController {
 
 	private Logger logger = LoggerFactory.getLogger(UserController.class);
+
+	@Autowired
+	User authUser;
 
 	@Autowired
 	public UserController(UserService userService) {
@@ -34,15 +32,23 @@ public class UserController {
 	@GetMapping(value = "users")
 	public ResponseEntity<List<User>> getAll(){
 
-		return new ResponseEntity<>(userService.findAll(), HttpStatus.OK);
-		
+		if (authUser.getProfile() != null && authUser.getProfile().equals("admin")) {
+			return new ResponseEntity<>(userService.findAll(), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
 	}
 	
 	@GetMapping(value = "users/{id}")
-	public ResponseEntity<User> getEquipmentById(@PathVariable("id") Integer id){
+	public ResponseEntity<User> getUserById(@PathVariable("id") Integer id){
 		
 		try {
-			return new ResponseEntity<>(userService.findById(id), HttpStatus.OK);
+			if ((authUser.getProfile() != null) &&
+					(authUser.getProfile().equals("admin") || Objects.equals(authUser.getId(), id))) {
+				return new ResponseEntity<>(userService.findById(id), HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			}
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		} catch (RuntimeException e) {
@@ -52,10 +58,14 @@ public class UserController {
 	}
 	
 	@PostMapping(value = "users")
-	public ResponseEntity<User> setEquipment(@RequestBody User user) {
+	public ResponseEntity<User> setUser(@RequestBody User user) {
 		
 		try {
-			return new ResponseEntity<>(userService.save(user), HttpStatus.OK);
+			if (authUser.getProfile() != null && authUser.getProfile().equals("admin")) {
+				return new ResponseEntity<>(userService.save(user), HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			}
 		} catch (RuntimeException e) {
 			logger.error("Error on creating user: " + e); 
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -64,11 +74,16 @@ public class UserController {
 	}
 	
 	@PutMapping(value = "users/{id}")
-	public ResponseEntity<User> updateEquipment(@RequestBody User user, @PathVariable("id") Integer id) {
+	public ResponseEntity<User> putUser(@RequestBody User user, @PathVariable("id") Integer id) {
 		
 		try {
-			user.setId(id);
-			return new ResponseEntity<>(userService.edit(user), HttpStatus.OK);
+			if ((authUser.getProfile() != null) &&
+					(authUser.getProfile().equals("admin") || Objects.equals(authUser.getId(), id))) {
+				user.setId(id);
+				return new ResponseEntity<>(userService.edit(user), HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			}
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);		
 			
@@ -77,13 +92,38 @@ public class UserController {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}		
 	}
-	
+
+	@PatchMapping(value = "users/{id}")
+	public ResponseEntity<User> patchUser(@RequestBody User user, @PathVariable("id") Integer id) {
+
+		try {
+			if (authUser.getProfile() != null &&
+					(authUser.getProfile().equals("admin") || Objects.equals(authUser.getId(), id))) {
+				user.setId(id);
+				return new ResponseEntity<>(userService.patch(user), HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			}
+		} catch (NoSuchElementException e) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+		} catch (RuntimeException e) {
+			logger.error("Error on updating user: " + e);
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+	}
+
 	@DeleteMapping(value = "users/{id}")
-	public ResponseEntity<User> deleteEquipment(@PathVariable("id") Integer id) {
+	public ResponseEntity<User> deleteUser(@PathVariable("id") Integer id) {
 		
 		try {
-			userService.delete(id);
-			return new ResponseEntity<>(HttpStatus.OK);			
+			if (authUser.getProfile() != null &&
+					(authUser.getProfile().equals("admin") || Objects.equals(authUser.getId(), id))) {
+				userService.delete(id);
+				return new ResponseEntity<>(HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			}
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
